shortened code by creating a parent class that has all methods needed
