+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Movable.java
@@ -1,16 +1,18 @@
 package org.firstinspires.ftc.teamcode;
 
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.hardware.Servo;
 
-@Autonomous
-public class DetestmentAndWoeFINAL extends LinearOpMode {
-    static private DcMotor FLW;
-    static private DcMotor BLW;
-    static private DcMotor FRW;
-    static private DcMotor BRW;
+abstract class Movable extends LinearOpMode {
+    static protected DcMotor FLW;
+    static protected DcMotor BLW;
+    static protected DcMotor FRW;
+    static protected DcMotor BRW;
+
+    //    static protected Servo ServoR;
+//    static protected Servo ServoL;
 
     @Override
     public void runOpMode() throws InterruptedException {
@@ -18,44 +20,22 @@ public class DetestmentAndWoeFINAL extends LinearOpMode {
         BLW = hardwareMap.get(DcMotor.class, "BLW");
         BRW = hardwareMap.get(DcMotor.class, "BRW");
         FRW = hardwareMap.get(DcMotor.class, "FRW");
+        ScissorLiftR = hardwareMap.get(DcMotor.class, "ScissorLiftR"); // change config if needed
+        ScissorLiftL = hardwareMap.get(DcMotor.class, "ScissorLiftL");
+//        ServoR = hardwareMap.get(Servo.class, "ServoR");
+//        ServoL = hardwareMap.get(Servo.class, "ServoL");
 
         telemetry.addData("Status", "Initialized");
         telemetry.update();
-        // Wait for the game to start (driver presses PLAY)
-        waitForStart();
-
-        // run until the end of the match (driver presses STOP)
-        // Example of controlling movements and then stopping after each step.
-        powerWheels(150, "right");
-        powerWheels(900, "forward");
-        turnRobot(200, "right"); // turn 180d
-        powerWheels(150, "left");
-
-        for (int i = 1; i <= 2; i++) {
-            powerWheels(800, "forward");
-            powerWheels(800, "backward");
-            powerWheels(150, "left");
-        }
-
-        powerWheels(800, "forward");
-        powerWheels(400, "backward");
-        wait(1000); // wait for the final specimen to be put in
-
-
-        // Stop motors after completing the parking routine
-        disablePower();
-        telemetry.addData("Status", "Completed");
-        telemetry.update();
     }
 
-
-
+    // front right wheel is the only one WRONG
     public void powerWheels(int miliseconds, String direction) {
         switch (direction) {
             case "forward":
                 FLW.setPower(0.3);
                 BLW.setPower(0.3);
-                FRW.setPower(0.3);
+                FRW.setPower(-0.3);
                 BRW.setPower(-0.3);
                 sleep(miliseconds);
                 disablePower();
@@ -63,7 +43,7 @@ public class DetestmentAndWoeFINAL extends LinearOpMode {
             case "backward":
                 FLW.setPower(-0.3);
                 BLW.setPower(-0.3);
-                FRW.setPower(-0.3);
+                FRW.setPower(0.3);
                 BRW.setPower(0.3);
                 sleep(miliseconds);
                 disablePower();
@@ -71,7 +51,7 @@ public class DetestmentAndWoeFINAL extends LinearOpMode {
             case "right":
                 FLW.setPower(0.3);
                 BLW.setPower(-0.3);
-                FRW.setPower(-0.3);
+                FRW.setPower(0.3);
                 BRW.setPower(-0.3);
                 sleep(miliseconds);
                 disablePower();
@@ -79,7 +59,7 @@ public class DetestmentAndWoeFINAL extends LinearOpMode {
             case "left":
                 FLW.setPower(-0.3);
                 BLW.setPower(0.3);
-                FRW.setPower(0.3);
+                FRW.setPower(-0.3);
                 BRW.setPower(0.3);
                 sleep(miliseconds);
                 disablePower();
@@ -92,7 +72,7 @@ public class DetestmentAndWoeFINAL extends LinearOpMode {
             case "left":
                 FLW.setPower(-0.3);
                 BLW.setPower(-0.3);
-                FRW.setPower(0.3);
+                FRW.setPower(-0.3);
                 BRW.setPower(-0.3);
                 sleep(miliseconds);
                 disablePower();
@@ -100,7 +80,7 @@ public class DetestmentAndWoeFINAL extends LinearOpMode {
             case "right":
                 FLW.setPower(0.3);
                 BLW.setPower(0.3);
-                FRW.setPower(-0.3);
+                FRW.setPower(0.3);
                 BRW.setPower(0.3);
                 sleep(miliseconds);
                 disablePower();
@@ -108,19 +88,46 @@ public class DetestmentAndWoeFINAL extends LinearOpMode {
         }
     }
 
-    public static void disablePower() {
+    public void disableScissorPower() {
+        ScissorLiftL.setPower(0);
+        ScissorLiftR.setPower(0);
+    }
+    public void disablePower() {
         FLW.setPower(0);
         BLW.setPower(0);
         FRW.setPower(0);
         BRW.setPower(0);
     }
 
-    public void updatePhoneConsole() {
-        telemetry.addData("FLW Power", FLW.getPower());
-        telemetry.addData("BLW Power", BLW.getPower());
-        telemetry.addData("FRW Power", FRW.getPower());
-        telemetry.addData("BRW Power", BRW.getPower());
-        telemetry.addData("Status", "Running");
-        telemetry.update();
+    public void powerScissorLift(int miliseconds, String direction) {
+        switch (direction) {
+            case "up":
+                ScissorLiftL.setPower(1);
+                ScissorLiftR.setPower(1);
+                sleep(miliseconds);
+                disableScissorPower();
+                break;
+            case "down":
+                ScissorLiftL.setPower(-1);
+                ScissorLiftR.setPower(-1);
+                sleep(miliseconds);
+                disableScissorPower();
+                break;
+        }
     }
+//
+//    public void claw(String toggle) {
+//        switch (toggle) {
+//            case "open":
+//                ServoR.setPosition(1);
+//                ServoL.setPosition(1);
+//                break;
+//            case "close":
+//                ServoR.setPosition(0);
+//                ServoL.setPosition(0);
+//                break;
+//        }
+//    }
+
+    abstract void updatePhoneConsole();
 }
